+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.3.0] - 2020-07-11
 ### Added
 - `extend_initialization_flow` alias method for `SmartCore::Initializer.ext_init`;
+- Access methods to the instance attribute lists (`#__params`, `#__options__`, `#__attributes__`);
+
+### Changed
+- Updated development dependencies;
 
 ## [0.2.0] - 2020-05-16
 ### Changed